pyproject.toml
[tool.poetry]
name = "be_centratura"
version = "0.1.0"
description = ""
authors = ["i-team <i-team@injenia.it>"]



[tool.poetry.dependencies]
python = "^3.7"

uvicorn = "^0.11.1"
fastapi = "^0.45.0"
pydantic = { version = "^1.3", extras = ["email"] }

passlib = {version = "^1.7", extras = ["bcrypt"]}

pyjwt = "^1.7"

asyncpg = "^0.20.0"
databases = { version = "^0.2.6", extras = ["postgresql"] }

python-slugify = "^4.0"
Unidecode = "^1.1"

loguru = "^0.4.0"
httptools = "^0.1.1"
aiocontextvars = "^0.2.2"
contextvars = "^2.4"
async-exit-stack = "^1.0.1"
async-generator = "^1.10"
python-multipart = "^0.0.5"
Pillow = "^8.2.0"
tensorflow = "2.3"
httpx = "0.16.0"


[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3"
autoflake = "^1.3"

wemake-python-styleguide = "^0.13.1"
mypy = "^0.740"
flake8-fixme = "^1.1"

pytest = "^5.3"
pytest-cov = "^2.8"
pytest-asyncio = "^0.14.0"
pytest-xdist = "^1.30"
docker = "^4.0"
asgi-lifespan = "^0.6.0"
jupyterlab = "^3.1.10"


[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"

